+++ src/layouts/Dashboard/index.tsx
@@ -3,6 +3,8 @@ import * as React from 'react';
 import styled from 'styled-components';
 import imageBg from '@/assets/images/Premark.png';
 import Sidebar from '@/components/Sidebar';
+import HeaderV2 from '@/components/HeaderV2';
+import { Box } from '@mui/material';
 
 export interface ILayoutProviderProps {
 	children: React.ReactNode;
@@ -12,7 +14,16 @@ export default function DashboardLayout({ children }: ILayoutProviderProps) {
 	return (
 		<LayoutWrapper>
 			<Sidebar />
-			<Container>{children}</Container>
+			<Box
+				position={'relative'}
+				marginLeft={'240px'}
+				width={'100%'}
+				paddingX={'32px'}
+				paddingY={'16px'}
+			>
+				<HeaderV2 />
+				<Box marginTop={'100px'}>{children}</Box>
+			</Box>
 		</LayoutWrapper>
 	);
 }
@@ -24,9 +35,4 @@ const LayoutWrapper = styled.div`
 	image-rendering: pixelated;
 	display: flex;
 	min-height: 100vh;
-`;
-const Container = styled.div`
-	padding: 20px 36px;
-	width: 100%;
-	margin-left: 240px;
 `;
